Add files via upload
Welcome to my image viewer application built with Python and Tkinter!

This is a simple yet powerful app that allows you to view and browse images in a variety of formats. Whether you're a photographer, designer, or just someone who loves visual media, this app is perfect for you.

Some of the key features of this app include:

Easy navigation through a variety of image formats, including JPEG, PNG, GIF, and more.
In future -> Simple controls for zooming, panning, and rotating images to get the perfect view.
In future -> Customizable preferences for how the app displays images, including size, quality, and more.
A sleek and intuitive interface that makes it easy to find and organize your images.
I built this app using Python and the Tkinter library, which allowed me to create a lightweight and responsive user interface that runs smoothly on any device. I'm constantly working on new features and improvements, so stay tuned for updates!

Feel free to download the code and try it out for yourself. If you have any feedback or suggestions, please don't hesitate to reach out. Thanks for checking out my image viewer app!